Tested and fixed most Example Notebooks
