+++ odp_sdk/utils/__init__.py
@@ -6,3 +6,9 @@ from .odp_geo import (
     grid_rect_members,
     index_rect_members
 )
+
+try:
+    import odp_sdk.utils.numeric as numeric
+    import odp_sdk.utils.visual as visual
+except ImportError:
+    pass
