+++ odp_sdk/utils/visual/__init__.py
@@ -1,0 +1,24 @@
+try:
+    import seaborn
+    import cartopy
+
+    del seaborn
+    del cartopy
+
+except ImportError as e:
+    raise ImportError("Missing dependencies for extended functionalities."
+                      "Please install package using the [all]-option.\n"
+                      "Example: pip install odp_sdk[all]\n"
+                      ">>>> " + str(e))
+
+from .casts import (
+    plot_casts,
+    plot_datasets,
+    plot_distributions,
+    plot_grid,
+    plot_meta_stats,
+    plot_nulls,
+    geo_map,
+    get_units,
+    missing_values
+)
