+++ tox.ini
@@ -11,7 +11,13 @@ commands = check-manifest
 description = Invoke sphinx-build to build the HTML docs
 deps =
     .[all]
+    sphinx
+    sphinx_rtd_theme
     -r docs/requirements-docs.txt
+extras = docs
+commands =
+    sphinx-build -d "{toxworkdir}/docs_doctree" docs/source "{toxworkdir}/docs_out" --color -W -bhtml {posargs}
+    python -c 'import pathlib; print("documentation available under file://\{0\}".format(pathlib.Path(r"{toxworkdir}") / "docs_out" / "index.html"))'
 
 [testenv:notebooks]
 deps =
